Update MacOS_source_build.md
adding json-c to the requirements. Although it did not work for me with configure option 1, the second option worked.

Updated the install folder for the proj4 as there is a document install which does not allow to make such folder.
Follow same structure for install as was done for GDAL.

Provide the absolute path for proj install.

The gdal build will pickup the path to python which could be python2.
Made it more explicit.