There are two unwrap stitching methods (overlap and 2stage). The latter is not yet functional. The different methods are implemented in a child stitching class, with mutual functionality in a parent class. Both child classes aim to build a mapping dictionary that tells per product, how much the phase of each connected component needs to be offset by an integer 2pi cycles, and also indicates what the new connected component number will be for the merged file. The connected components of individual products are incremented in the merged connected component, but where connected component zero has been concatenated.
Intermediate files are being generated in temp folder and then at the final __ createImages__ step they are warped into the new images. For connected component the temp file is a tiff file, for unwrapped the temp files are an integer tiff file, a vrt scaling the tiff with 2pi, a vrt with the sum of the original file and the scaling. The VRT files contain proj and transformation information.

The commandline parse has been included to be inline with the other functions. An example on how the call the functionality within python can be found in the main.

Currently it still has some limitations that need to be further addressed:

    Cropping based on BBOX and cutline options. To do this all at the same gdalwarp step might require the tiff files to be updated to incorporate the proje and transf too.
    Update the gdalwarp at the file-generation step (createImages) to be pythonic
    When we extract a single product the “overlap” method could default to a vrt?
    The 2stage method still needs work and debugging, default is set for overlap method

Below you can see an example over Hawaii before and after running the code. You can see a small edge effect using the overlap method. Would be good if this can be tested on other locations for the 2-stager

To run the code

unwrapStitching.py -f "*.nc"