Fix bug when extracting iono and tropo layer
Addressing issue #420 , where an error is raised when just either the ionosphere and tropo layer are extracted because the dimensions were not properly passed and accounted for.

Specifically, the program checks for internal consistency between the dimensions of all output layers. This information was not initialized and passed for the ionosphere and tropo layers, so when this check is executed the program fails.